command-service/requirements.txt
fastapi==0.103.0
uvicorn==0.20.0
dataclasses-json==0.5.7
# pydantic==1.10.5
pyyaml==6.0.1
urllib3==2.0.7
# protobuf==3.20.*
# pyhelm==2.14.5
# kubernetes==28.1.0
# helm_operator_sdk
# psycopg2==2.9.5
psycopg2-binary
dacite==1.8.0
backoff==2.2.1
tenacity==8.2.2
kafka-python-ng==2.2.2
boto3
prometheus-client
confluent-kafka==2.5.3
requests